Separate share token tests into a new spec
